requirements.txt
streamlit
gitpython
google-generativeai